% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_json_to_excel.R
\name{question_json_to_excel}
\alias{question_json_to_excel}
\title{Question-metadata json to excel}
\usage{
question_json_to_excel(in_path, out_file)
}
\arguments{
\item{in_path}{Input path, e.g. "./input/questions/ins1"}

\item{out_file}{Path to the excel file to be generated, e.g.
"./output/projectname.xlsx"}
}
\description{
This script generates a question metadata excel file from question metadata
json files.
}
