% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_jsons_to_excel.R
\name{question_jsons_to_excel}
\alias{question_jsons_to_excel}
\title{Question-metadata jsons to excel}
\usage{
question_jsons_to_excel(in_path, out_file, instrument_number)
}
\arguments{
\item{in_path}{Input path, e.g. "./input/questions/ins1"}

\item{out_file}{Path of the excel file to be generated, e.g.
"./output/projectname.xlsx"}

\item{instrument_number}{Numeric}
}
\description{
This script generates a question metadata excel file from question metadata
json files.
}
