% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_trim_excel.R
\name{read_and_trim_excel}
\alias{read_and_trim_excel}
\title{Read excel files and get rid of unwanted whitespace}
\usage{
read_and_trim_excel(path, sheet)
}
\arguments{
\item{path}{the path to the excel file}

\item{sheet}{the sheet that should be edited}
}
\description{
Read excel files and get rid of unwanted whitespace
}
