% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort-images.R
\name{sort_images}
\alias{sort_images}
\title{SORT IMAGES IN DIRECTORY STRUCTURE}
\usage{
sort_images(pathInImages, pathOutImages, pathXlsxFile)
}
\arguments{
\item{pathInImages}{path containing folders ins1 ins2 etc.}

\item{pathOutImages}{path to newly generated json files}

\item{pathXlsxFile}{path to the excel file}
}
\description{
Dieses Skript unterstützt bei der Einsortierung der Fragenbilder (pngs) in die Ordnerstruktur,
die für den Upload ins Metadatensystem benötigt wird.
Das Skript ist auszuführen, nachdem die json Dateien erzeugt wurden und die Ordnerstruktur angelegt wurde.
Angepasst werden muss der Pfad an dem die Bilder liegen, der Pfad an dem die Bilder gespeichert werden
sollen und der Pfad an dem die Question Excel Tabelle liegt (+ evtl. Projektname).
Das Tabellenblatt "images" der Exceltabelle muss vorher ausgefüllt werden. Dann ist es auch m?glich, wenn mehrere Bilder pro Frage vorliegen,
diese richtig einzusortieren.
Ordnerstruktur der Bilder vor dem Einsortieren:
|--ins1
    |--1.1_1.png
    |--1.1_2.png
|--ins2
    |--1.1.png
    |--1.2.png
Ordnerstruktur nach dem Einsortieren:
|--ins1
    |--images
        |--1.1
            |--1.1_1.png
            |--1.1_2.png
    |--1.1.json
    |--1.2.json
}
\details{
|--ins2
    |--images
        |--1.1
            |--1.1.png
        |--1.2
            |--1.2.png
    |--1.1.json
    |--1.2.json
}
