% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_multiple_leading_trailing_ws.R
\name{trim_multiple_leading_trailing_ws}
\alias{trim_multiple_leading_trailing_ws}
\title{Trim multiple leading and trailing whitespace}
\usage{
trim_multiple_leading_trailing_ws(x)
}
\arguments{
\item{x}{the object whose whitespace should be trimmed}
}
\description{
Trim multiple leading and trailing whitespace
}
\keyword{internal}
